requirements.txt
Flask
requests
beautifulsoup4
python-dotenv
html5lib
Selenium
gunicorn